backend/requirements-pinecone.txt
# Simple requirements for Pinecone knowledge base backend
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
firebase-admin==6.4.0
requests==2.31.0
httpx==0.25.2
langchain==0.2.12
langchain-community==0.2.12
langchain-openai==0.1.8
pinecone-client==2.2.4
openai==1.3.7
tiktoken==0.5.2
PyPDF2==3.0.1
pdfplumber==0.10.3
beautifulsoup4==4.12.2
lxml==4.9.3
